report

+++ ConsultationApi/src/main/java/com/consultationapi/consultationapi/data/user/ReadUser.java
@@ -26,12 +26,12 @@ public class ReadUser  extends ConnectionAttributes {
                 user.setUserName(r.getString(3));
                 user.setPassword(r.getString(4));
                 user.setEmail(r.getString(5));
-              //  user.setDate(r.getDate(6));
+                //user.setDate(r.getDate(6));
                 user.setType(r.getString(7));
                 user.setSaldo( BigDecimal.valueOf(r.getDouble(8)));
                 user.setAddress(r.getString(10));
-                user.setPhone(r.getInt(11));
-                user.setCui(r.getInt(12));
+                user.setPhone(r.getLong(11));
+                user.setCui(r.getLong(12));
                 typeUser=user;
             }
         }catch (SQLException ex) {
